tests/wpt/web-platform-tests/trusted-types/Location-assign.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  test(t => {
    let p = createURL_policy(window, 1);
    let url = p.createURL(location.href + "#xxx");
    location.assign(url);
    assert_equals("" + url, location.href, "location href");
  }, "location.assign via policy (successful URL transformation).");
</script>
